requirements.txt
google-api-python-client==2.161.0
python-dateutil==2.9.0.post0
pandas==2.2.3
tqdm==4.67.1
pymongo==4.11.1
openpyxl==3.1.5
mysql-connector-python==9.2.0
google-auth-oauthlib==1.2.1
insta-scrape==2.1.2
selenium==4.30.0
python-dotenv==1.0.1